HS-190 | Update music controller to change music depending on day part

+++ core/src/main/java/pl/agh/edu/actor/slider/PercentSliderComponent.java
@@ -3,16 +3,16 @@ package pl.agh.edu.actor.slider;
 import java.util.function.Function;
 
 public class PercentSliderComponent extends SliderComponent {
-	private final Function<Long, Void> stateChangeHandler;
+	private final Function<Float, Void> stateChangeHandler;
 
-	public PercentSliderComponent(String name, Function<Long, Void> stateChangeHandler) {
+	public PercentSliderComponent(String name, Function<Float, Void> stateChangeHandler) {
 		super(name, "%", 0f, 100f, 1f);
 		this.stateChangeHandler = stateChangeHandler;
 		valueLabel.setText(getPercentageValue() + " " + suffix);
 	}
 
-	private long getPercentageValue() {
-		return (long) this.getValue();
+	private float getPercentageValue() {
+		return this.getValue();
 	}
 
 	@Override
